Refactor project structure and update examples
Reorganized project from "samples" to "examples", updating references, paths, and documentation. Removed old sample-related files and added new example-related files, including controllers, services, and configurations. Updated localization resources and project setup for .NET 8.0.
