+++ OperationResults.sln
@@ -11,17 +11,17 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		src\Directory.Build.props = src\Directory.Build.props
 	EndProjectSection
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{40EA0531-BCD9-4D33-9114-D1BB9EA7DDBA}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "examples", "examples", "{40EA0531-BCD9-4D33-9114-D1BB9EA7DDBA}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OperationResults.Sample", "samples\Controllers\OperationResults.Sample\OperationResults.Sample.csproj", "{11C60254-4212-42ED-ACDF-CAC830E8BEB8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OperationResults.Example", "examples\Controllers\OperationResults.Example\OperationResults.Example.csproj", "{11C60254-4212-42ED-ACDF-CAC830E8BEB8}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OperationResults.AspNetCore", "src\OperationResults.AspNetCore\OperationResults.AspNetCore.csproj", "{375F6E50-F06A-4832-A712-D2C7A195DE64}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OperationResults.Sample.BusinessLayer", "samples\OperationResults.Sample.BusinessLayer\OperationResults.Sample.BusinessLayer.csproj", "{2F6D04EA-8B0C-44E3-A483-CC02D376A817}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OperationResults.Example.BusinessLayer", "examples\OperationResults.Example.BusinessLayer\OperationResults.Example.BusinessLayer.csproj", "{2F6D04EA-8B0C-44E3-A483-CC02D376A817}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OperationResults.Sample.DataAccessLayer", "samples\OperationResults.Sample.DataAccessLayer\OperationResults.Sample.DataAccessLayer.csproj", "{0336F79E-CB4A-435C-A1DE-AD2E8621F106}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OperationResults.Example.DataAccessLayer", "examples\OperationResults.Example.DataAccessLayer\OperationResults.Example.DataAccessLayer.csproj", "{0336F79E-CB4A-435C-A1DE-AD2E8621F106}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OperationResults.Sample.Shared", "samples\OperationResults.Sample.Shared\OperationResults.Sample.Shared.csproj", "{A6935B6E-0E91-4B72-9B0D-083918292998}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OperationResults.Example.Shared", "examples\OperationResults.Example.Shared\OperationResults.Example.Shared.csproj", "{A6935B6E-0E91-4B72-9B0D-083918292998}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Controllers", "Controllers", "{6A120E62-C0F2-487F-BC61-D29FCD42F8ED}"
 EndProject
@@ -29,7 +29,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OperationResults.AspNetCore
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Minimal APIs", "Minimal APIs", "{085A97A4-3BFB-458C-868F-8B9437260664}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OperationResults.Sample", "samples\MinimalApis\OperationResults.Sample\OperationResults.Sample.csproj", "{15965E02-9802-42B1-9CA5-7E5A84395477}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OperationResults.Example", "examples\MinimalApis\OperationResults.Example\OperationResults.Example.csproj", "{15965E02-9802-42B1-9CA5-7E5A84395477}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
