+++ README.md
@@ -26,7 +26,7 @@ _Note: This is the library to use if you're working with Controller._
 
 This library provides HttpContext extension methods to automatically map Operation Results (that may come, for example, from a business layer) to HTTP responses, along with the appropriate status codes.
 
-A full example is available in the [Samples](https://github.com/marcominerva/OperationResults/tree/master/samples) folder. Search for the registration in the [Program.cs](https://github.com/marcominerva/OperationResults/blob/master/samples/Controllers/OperationResults.Sample/Program.cs#L24-L40) file and the usage in [Controllers](https://github.com/marcominerva/OperationResults/tree/master/samples/Controllers/OperationResults.Sample/Controllers) folder.
+A full example is available in the [Example folder](https://github.com/marcominerva/OperationResults/tree/master/examples). Search for the registration in the [Program.cs](https://github.com/marcominerva/OperationResults/blob/master/examples/Controllers/OperationResults.Example/Program.cs#L24-L40) file and the usage in [Controllers folder](https://github.com/marcominerva/OperationResults/tree/master/examples/Controllers/OperationResults.Example/Controllers).
 
 **Installation**
 
@@ -43,7 +43,7 @@ _Note: This is the library to use if you're working with Minimal APIs._
 
 This library provides HttpContext extension methods to automatically map Operation Results (that may come, for example, from a business layer) to HTTP responses, along with the appropriate status codes.
 
-A full example is available in the [Samples](https://github.com/marcominerva/OperationResults/tree/master/samples) folder. Search for the [registration](https://github.com/marcominerva/OperationResults/blob/master/samples/MinimalApis/OperationResults.Sample/Program.cs#L23-L35) and the [usage](https://github.com/marcominerva/OperationResults/blob/master/samples/MinimalApis/OperationResults.Sample/Program.cs#L51-L106) in [Program.cs](https://github.com/marcominerva/OperationResults/blob/master/samples/MinimalApis/OperationResults.Sample/Program.cs) file.
+A full example is available in the [Examples folder](https://github.com/marcominerva/OperationResults/tree/master/samples). Search for the [registration](https://github.com/marcominerva/OperationResults/blob/master/examples/MinimalApis/OperationResults.Example/Program.cs#L23-L35) and the [usage](https://github.com/marcominerva/OperationResults/blob/master/examples/MinimalApis/OperationResults.Example/Program.cs#L51-L106) in [Program.cs](https://github.com/marcominerva/OperationResults/blob/master/examples/MinimalApis/OperationResults.Example/Program.cs) file.
 
 **Installation**
