+++ examples/Controllers/OperationResults.Example/Resources/Messages.resx
