+++ examples/MinimalApis/OperationResults.Example/appsettings.Development.json
