+++ examples/OperationResults.Example.BusinessLayer/CustomFailureReasons.cs
@@ -1,0 +1,6 @@
+﻿namespace OperationResults.Example.BusinessLayer;
+
+public static class CustomFailureReasons
+{
+    public const int NotAvailable = FailureReasons.GenericError + 1;
+}
