+++ examples/OperationResults.Example.BusinessLayer/OperationResults.Example.BusinessLayer.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\OperationResults\OperationResults.csproj" />
+    <ProjectReference Include="..\OperationResults.Example.DataAccessLayer\OperationResults.Example.DataAccessLayer.csproj" />
+    <ProjectReference Include="..\OperationResults.Example.Shared\OperationResults.Example.Shared.csproj" />
+  </ItemGroup>
+
+</Project>
