+++ examples/OperationResults.Example.BusinessLayer/Services/Interfaces/IImageService.cs
@@ -1,0 +1,6 @@
+﻿namespace OperationResults.Example.BusinessLayer.Services.Interfaces;
+
+public interface IImageService
+{
+    Task<Result<ByteArrayFileContent>> GetImageAsync();
+}
