+++ examples/OperationResults.Example.DataAccessLayer/Entities/Person.cs
@@ -1,0 +1,16 @@
+﻿namespace OperationResults.Example.DataAccessLayer.Entities;
+
+public class Person
+{
+    public Guid Id { get; set; }
+
+    public string FirstName { get; set; }
+
+    public string LastName { get; set; }
+
+    public string Email { get; set; }
+
+    public string City { get; set; }
+
+    public DateTime CreateDate { get; set; }
+}
