+++ examples/OperationResults.Example.Shared/Models/PersonWithImage.cs
@@ -1,0 +1,8 @@
+namespace OperationResults.Example.Shared.Models;
+
+public class PersonWithImage
+{
+    public Person Person { get; set; }
+
+    public byte[] Image { get; set; }
+}
