+++ examples/OperationResults.Example.Shared/OperationResults.Example.Shared.csproj
