--- samples/Controllers/OperationResults.Sample/OperationResults.Sample.csproj
@@ -1,34 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="8.0.8" />
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="6.8.0" />
-        <PackageReference Include="TinyHelpers.AspNetCore" Version="3.1.9" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\..\src\OperationResults.AspNetCore\OperationResults.AspNetCore.csproj" />
-      <ProjectReference Include="..\..\OperationResults.Sample.BusinessLayer\OperationResults.Sample.BusinessLayer.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Compile Update="Resources\Messages.Designer.cs">
-        <DesignTime>True</DesignTime>
-        <AutoGen>True</AutoGen>
-        <DependentUpon>Messages.resx</DependentUpon>
-      </Compile>
-    </ItemGroup>
-
-    <ItemGroup>
-      <EmbeddedResource Update="Resources\Messages.resx">
-        <Generator>PublicResXFileCodeGenerator</Generator>
-        <LastGenOutput>Messages.Designer.cs</LastGenOutput>
-      </EmbeddedResource>
-    </ItemGroup>
-
-</Project>
