--- samples/OperationResults.Sample.BusinessLayer/CustomFailureReasons.cs
@@ -1,6 +1,0 @@
-﻿namespace OperationResults.Sample.BusinessLayer;
-
-public static class CustomFailureReasons
-{
-    public const int NotAvailable = FailureReasons.GenericError + 1;
-}
