--- samples/OperationResults.Sample.BusinessLayer/OperationResults.Sample.BusinessLayer.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\OperationResults\OperationResults.csproj" />
-    <ProjectReference Include="..\OperationResults.Sample.DataAccessLayer\OperationResults.Sample.DataAccessLayer.csproj" />
-    <ProjectReference Include="..\OperationResults.Sample.Shared\OperationResults.Sample.Shared.csproj" />
-  </ItemGroup>
-
-</Project>
