--- samples/OperationResults.Sample.BusinessLayer/Services/Interfaces/IImageService.cs
@@ -1,6 +1,0 @@
-﻿namespace OperationResults.Sample.BusinessLayer.Services.Interfaces;
-
-public interface IImageService
-{
-    Task<Result<ByteArrayFileContent>> GetImageAsync();
-}
