--- samples/OperationResults.Sample.Shared/Models/PersonWithImage.cs
@@ -1,8 +1,0 @@
-namespace OperationResults.Sample.Shared.Models;
-
-public class PersonWithImage
-{
-    public Person Person { get; set; }
-
-    public byte[] Image { get; set; }
-}
